The Explorer
Bai Vylcho is very an enthusiastic explorer. His passion are the
diamonds, he just adores them. Today he is going on an expedition to
collect all kind of diamonds, no matter small or large. Help your friend
to find all the diamonds in the biggest known cave "The Console Cave".
At the only input line you will be given the width of the diamond. The
char that forms the outline of the diamonds is '*' and the surrounding
parts are made of '-' (see the examples). Your task is to print a
diamond of given size n.